+++ Changelog.md
@@ -1,5 +1,6 @@
 ## Fixed
 - Contraptions can be placed on a command block that transforming into a lily pad.
+- Instant evoke function now no longer can be used on nocturnal contraptions during day.
 
 ---